Change file schema -> lectureFile and move foreign key to lectureFile.
To avoid conflicts with the file keyword in SQL the file schema was
changed to lectureFile. The name change should also be more clearly identify what the files
are used for.

Since a single lecture can have multiple files associated with it, the
foreign key was moved into the lectureFile schema.