Fix bug where home page was sending requests for courses in infinite loop.
When a playlist with no courses was clicked on then the client would attempt
to request for a list of courses for that playlist.  Since the playlist
had no courses the client would attempt again to get more courses and
repeat this process in an infinite loop.

Removing the componentDidUpdate function, which was no longer serving a purpose, fixes this issue.