Fix bug where dropdown was not working for course and playlists.
The settings button for course and playlist would sometimes stop working
and the dropdown would not display the items in its list. This bug
occured when a React course or playlist component would refresh, but the
jquery initializing the dropdown to work properly was not run again.

To fix this the jquery that controlls the dropdown was moved out of the
home component and into the onClick handlers inside the playlist and
course components so that whenever a component was refreshed the onClick
handler would be properly initialized.